+++ src/ts/dom/dom.ts
@@ -4,7 +4,7 @@
  * A JavaScript library for converting between HTML and JSON, with binding, templating, attributes, and CSS support.
  * 
  * @file        dom.ts
- * @version     v2.1.0
+ * @version     v2.2.0
  * @author      Bunoon
  * @license     MIT License
  * @copyright   Bunoon 2024
@@ -16,12 +16,17 @@ import { Char } from "../data/enum";
 
 export namespace DomElement {
     export function create( container: HTMLElement, type: string ) : HTMLElement {
+        let result: any = createWithNoContainer( type );
+        container.appendChild( result );
+
+        return result;
+    }
+
+    export function createWithNoContainer( type: string ) : HTMLElement {
         const nodeType: string = type.toLowerCase();
         const isText: boolean = nodeType === "text";
 
         let result: any = isText ? document.createTextNode( Char.empty ) : document.createElement( nodeType );
-
-        container.appendChild( result );
 
         return result;
     }