+++ test/js/scripts.js
@@ -2,6 +2,7 @@ var _BODY = null;
 
 var _JSON_TEXTAREA = null,
     _INCLUDE_ATTRIBUTES = null,
+    _INCLUDE_DATA_ATTRIBUTES = null,
     _INCLUDE_CSS_STYLES = null,
     _INCLUDE_TEXT = null,
     _INCLUDE_CHILDREN = null,
@@ -14,12 +15,14 @@ var _JSON_TEXTAREA = null,
     _GENERATE_UNIQUE_MISSING_NAMES = null;
 
 var _SET_TEMPLATE_DATA = null,
-    _REMOVE_ATTRIBUTES = null,
-    _CLEAR_HTML = null,
+    _REMOVE_ORIGINAL_ATTRIBUTES = null,
+    _REMOVE_ORIGINAL_DATA_ATTRIBUTES = null,
+    _CLEAR_ORIGINAL_HTML = null,
     _WRITE_CSS_TO_HEAD = null,
     _CLEAR_CSS_FROM_HEAD = null,
     _LOG_TEMPLATE_DATA_WARNINGS = null,
     _ADD_ATTRIBUTES = null,
+    _ADD_DATA_ATTRIBUTES = null,
     _ADD_CSS_PROPERTIES = null,
     _ADD_TEXT = null,
     _ADD_CHILDREN = null,
@@ -34,6 +37,7 @@ var _SET_TEMPLATE_DATA = null,
 
         _JSON_TEXTAREA = document.getElementById( "json" );
         _INCLUDE_ATTRIBUTES = document.getElementById( "includeAttributes" );
+        _INCLUDE_DATA_ATTRIBUTES = document.getElementById( "includeDataAttributes" );
         _INCLUDE_CSS_STYLES = document.getElementById( "includeCssProperties" );
         _INCLUDE_TEXT = document.getElementById( "includeText" );
         _INCLUDE_CHILDREN = document.getElementById( "includeChildren" );
@@ -46,12 +50,14 @@ var _SET_TEMPLATE_DATA = null,
         _GENERATE_UNIQUE_MISSING_NAMES = document.getElementById( "generateUniqueMissingNames" );
 
         _SET_TEMPLATE_DATA = document.getElementById( "setTemplateData" );
-        _REMOVE_ATTRIBUTES = document.getElementById( "removeOriginalAttributes" );
-        _CLEAR_HTML = document.getElementById( "clearOriginalHTML" );
+        _REMOVE_ORIGINAL_ATTRIBUTES = document.getElementById( "removeOriginalAttributes" );
+        _REMOVE_ORIGINAL_DATA_ATTRIBUTES = document.getElementById( "removeOriginalDataAttributes" );
+        _CLEAR_ORIGINAL_HTML = document.getElementById( "clearOriginalHTML" );
         _WRITE_CSS_TO_HEAD = document.getElementById( "addCssToHead" );
         _CLEAR_CSS_FROM_HEAD = document.getElementById( "clearCssFromHead" );
         _LOG_TEMPLATE_DATA_WARNINGS = document.getElementById( "logTemplateDataWarnings" );
         _ADD_ATTRIBUTES = document.getElementById( "addAttributes" );
+        _ADD_DATA_ATTRIBUTES = document.getElementById( "addDataAttributes" );
         _ADD_CSS_PROPERTIES = document.getElementById( "addCssProperties" );
         _ADD_TEXT = document.getElementById( "addText" );
         _ADD_CHILDREN = document.getElementById( "addChildren" );
@@ -61,7 +67,7 @@ var _SET_TEMPLATE_DATA = null,
 
 function setupConfiguration() {
     $jhson.setConfiguration( {
-        nodeTypesToIgnore: [ "div" ]
+        safeMode: false
     } );
 }
 
@@ -69,6 +75,7 @@ function getJSONData() {
     _JSON_TEXTAREA.value = $jhson
         .json()
         .includeAttributes( _INCLUDE_ATTRIBUTES.checked )
+        .includeDataAttributes( _INCLUDE_DATA_ATTRIBUTES.checked )
         .includeCssProperties( _INCLUDE_CSS_STYLES.checked )
         .includeText( _INCLUDE_TEXT.checked )
         .includeChildren( _INCLUDE_CHILDREN.checked )
@@ -96,7 +103,15 @@ function getHTMLVariables() {
     console.log( variablesString );
 }
 
-function writeJSONData() {
+function writeJsonToHtmlData() {
+    getHtmlScope().write( _BODY );
+}
+
+function getJsonToHtmlElement() {
+    console.log( getHtmlScope().get() );
+}
+
+function getHtmlScope() {
     var htmlScope = $jhson
         .html()
         .json( _JSON_TEXTAREA.value );
@@ -110,15 +125,18 @@ function writeJSONData() {
     }
 
     htmlScope
-        .removeOriginalAttributes( _REMOVE_ATTRIBUTES.checked )
-        .clearOriginalHTML( _CLEAR_HTML.checked )
+        .removeOriginalAttributes( _REMOVE_ORIGINAL_ATTRIBUTES.checked )
+        .removeOriginalDataAttributes( _REMOVE_ORIGINAL_DATA_ATTRIBUTES.checked )
+        .clearOriginalHTML( _CLEAR_ORIGINAL_HTML.checked )
         .addCssToHead( _WRITE_CSS_TO_HEAD.checked )
         .clearCssFromHead( _CLEAR_CSS_FROM_HEAD.checked )
         .logTemplateDataWarnings( _LOG_TEMPLATE_DATA_WARNINGS.checked )
         .addAttributes( _ADD_ATTRIBUTES.checked )
+        .addDataAttributes( _ADD_DATA_ATTRIBUTES.checked )
         .addCssProperties( _ADD_CSS_PROPERTIES.checked )
         .addText( _ADD_TEXT.checked )
         .addChildren( _ADD_CHILDREN.checked )
-        .insertBefore( _INSERT_BEFORE.checked )
-        .write( _BODY );
+        .insertBefore( _INSERT_BEFORE.checked );
+
+    return htmlScope;
 }